plexmind/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
plexapi>=4.15.0
pydantic>=2.7.0
apscheduler>=3.10.4
python-multipart>=0.0.9
